Decompose MainWindow partials to reduce file size
